unitRation int -> double

+++ src/main/java/org/glo/giftw/domain/strategy/Field.java
@@ -9,7 +9,7 @@ public class Field implements Serializable
     public static final long serialVersionUID = 1L;
 
     private Vector dimensions;
-    private int unitRatio;  //nombre d'unités par mètre
+    private double unitRatio;  //nombre d'unités par mètre
     private String imagePath;
 
     public Field()
@@ -22,7 +22,7 @@ public class Field implements Serializable
         this(dimensions, 100, fieldImagePath);
     }
     
-    public Field(Vector dimensions, int unitRatio, String fieldImagePath)
+    public Field(Vector dimensions, double unitRatio, String fieldImagePath)
     {
         this.dimensions = dimensions;
         this.unitRatio = unitRatio;
@@ -39,12 +39,12 @@ public class Field implements Serializable
         this.dimensions = dimensions;
     }
     
-    public int getUnitRatio()
+    public double getUnitRatio()
     {
         return this.unitRatio;
     }
     
-    public void setUnitRatio(int unitRatio)
+    public void setUnitRatio(double unitRatio)
     {
         this.unitRatio = unitRatio;
     }
@@ -84,7 +84,7 @@ public class Field implements Serializable
     @Override
     public String toString()
     {
-        return "Dimension: " + this.dimensions.toString() + "\nUnit Ratio: " + Integer.toString(this.unitRatio) +
+        return "Dimension: " + this.dimensions.toString() + "\nUnit Ratio: " + Double.toString(this.unitRatio) +
                 "\nImage path: " + this.imagePath;
     }
